Cast chrono counts to long long in %lld LOG_TRACE calls
Addresses three Copilot review comments (TransmissionPolicyManager.cpp:119,
202, 266). This PR changed these LOG_TRACE format strings from %d to %lld but
passed std::chrono::milliseconds::rep directly. That rep is implementation-
defined and is long on LP64 (Linux/macOS), so %lld (which expects long long)
is a -Wformat mismatch -- an error under the project's -Wall -Werror in
logging-enabled (HAVE_MAT_LOGGING) builds, and formally UB in the varargs call.

Cast each count() to long long so the format always matches on every data model.
This mirrors the cast this PR already applies to delta (static_cast<unsigned
long long> with %llu) a few lines up.

Verified: clang 18 -Wall -Werror -Wextra flags the uncast %lld as
"format specifies type 'long long' but the argument has type 'rep' (aka 'long')"
and accepts the cast form. TransmissionPolicyManagerTests (40) pass.

+++ lib/tpm/TransmissionPolicyManager.cpp
@@ -116,7 +116,7 @@ namespace MAT_NS_BEGIN {
             if (delay.count() < 0 || m_timerdelay.count() < 0)
             {
                 LOG_TRACE("Negative delay(%lld) or m_timerdelay(%lld), no upload",
-                    delay.count(), m_timerdelay.count());
+                    static_cast<long long>(delay.count()), static_cast<long long>(m_timerdelay.count()));
                 return true;
             }
             if (m_scheduledUploadAborted)
@@ -199,7 +199,7 @@ namespace MAT_NS_BEGIN {
             m_isUploadScheduled = true;
             m_scheduledUploadTime = PAL::getMonotonicTimeMs() + delay.count();
             m_runningLatency = latency;
-            LOG_TRACE("SCHED upload %lld ms for lat=%d", delay.count(), m_runningLatency);
+            LOG_TRACE("SCHED upload %lld ms for lat=%d", static_cast<long long>(delay.count()), m_runningLatency);
             m_scheduledUpload = PAL::scheduleTask(&m_taskDispatcher, static_cast<unsigned>(delay.count()), this, &TransmissionPolicyManager::uploadAsync, latency);
         }
     }
@@ -263,7 +263,7 @@ namespace MAT_NS_BEGIN {
         // Rescheduling upload
         if (nextUpload.count() >= 0)
         {
-            LOG_TRACE("Scheduling upload in %lld ms", nextUpload.count());
+            LOG_TRACE("Scheduling upload in %lld ms", static_cast<long long>(nextUpload.count()));
             EventLatency proposed = calculateNewPriority();
             scheduleUpload(nextUpload, proposed); // reschedule uploadAsync again
         }
